src/main/java/com/mycompany/pojo/Object.java
/*
 * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
 * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
 */
package com.mycompany.pojo;

/**
 *
 * @author HP
 */
public class Object {
    private int id;
    private String name;

    public Object(int id, String name){
        this.id = id;
        this.name = name;
    }
    public Object(String name){
        this.name = name;
    }
    public Object(){
        
    }
    @Override
    public String toString() {
        return this.name;
    }    
    /**
     * @return the id
     */
    public int getId() {
        return id;
    }

    /**
     * @param id the id to set
     */
    public void setId(int id) {
        this.id = id;
    }

    /**
     * @return the name
     */
    public String getName() {
        return name;
    }

    /**
     * @param name the name to set
     */
    public void setName(String name) {
        this.name = name;
    }
    
}
